requirements.txt
streamlit 
plotly.express
streamlit_option_menu
PIL